# refactor: remove clone bound from entry()