--- a/foyer-storage/src/flusher.rs
+++ b/foyer-storage/src/flusher.rs
@@ -195,6 +195,7 @@ where
         {
             bytes += len;
             let index = Index::Region {
+                key: key.clone(),
                 view: self.region_manager.region(&region).view(offset as u32, len as u32),
             };
             let item = Item::new(sequence, index);