--- a/foyer-storage/src/generic.rs
+++ b/foyer-storage/src/generic.rs
@@ -361,10 +361,10 @@ where
         let inner = self.inner.clone();
         let now = Instant::now();
 
-        let index_value = inner.catalog.get(key);
+        let res = inner.catalog.get(key);
 
         async move {
-            let (_sequence, index) = match index_value {
+            let (_sequence, index) = match res {
                 Some(item) => item.consume(),
                 None => {
                     inner.metrics.op_duration_get_miss.observe(now.elapsed().as_secs_f64());
@@ -380,7 +380,7 @@ where
 
                     Ok(Some(CachedEntry::Shared { key, value }))
                 }
-                crate::catalog::Index::Region { view } => {
+                crate::catalog::Index::Region { key, view } => {
                     let region = view.id();
 
                     inner.region_manager.record_access(region);
@@ -391,7 +391,7 @@ where
                         Some(buf) => buf,
                         None => {
                             // Remove index if the storage layer fails to get it (because of region version mismatch).
-                            // inner.catalog.remove(index.key());
+                            inner.catalog.remove::<K>(key.as_ref());
                             inner.metrics.op_duration_get_miss.observe(now.elapsed().as_secs_f64());
                             return Ok(None);
                         }
@@ -407,7 +407,7 @@ where
                         }
                         Err(e) => {
                             // Remove index if the storage layer fails to get it (because of entry magic mismatch).
-                            // inner.catalog.remove(key);
+                            inner.catalog.remove::<K>(key.as_ref());
                             Err(e)
                         }
                     };
@@ -508,7 +508,7 @@ where
         let res = if let Some(mut iter) = RegionEntryIter::<K, V, D>::open(region).await? {
             while let Some((key, item)) = iter.next().await? {
                 sequence = std::cmp::max(sequence, *item.sequence());
-                catalog.insert(Arc::new(key), item);
+                catalog.insert(key, item);
             }
             region_manager.eviction_push(region_id);
             Some(sequence)
@@ -861,7 +861,7 @@ where
         }))
     }
 
-    pub async fn next(&mut self) -> Result<Option<(K, Item<K, V>)>> {
+    pub async fn next(&mut self) -> Result<Option<(Arc<K>, Item<K, V>)>> {
         let region_size = self.region.device().region_size();
         let align = self.region.device().align();
 
@@ -898,11 +898,11 @@ where
             let rel_start = EntryHeader::serialized_len() + header.value_len as usize;
             let rel_end = rel_start + header.key_len as usize;
 
-            let Ok(key) = bincode::deserialize_from(&slice.as_ref()[rel_start..rel_end]) else {
+            let Ok(key) = bincode::deserialize_from::<_, K>(&slice.as_ref()[rel_start..rel_end]) else {
                 return Ok(None);
             };
             drop(slice);
-            key
+            Arc::new(key)
         } else {
             drop(slice);
             let Some(s) = self.region.load_range(align_start..align_end).await? else {
@@ -921,6 +921,7 @@ where
         let info = Item::new(
             header.sequence,
             Index::Region {
+                key: key.clone(),
                 view: self.region.view(self.cursor as u32, entry_len as u32),
             },
         );
@@ -937,7 +938,7 @@ where
             Err(e) => return Err(e),
         };
 
-        let Index::Region { view } = item.index() else {
+        let Index::Region { key: _, view } = item.index() else {
             unreachable!("kv loaded from region must have index of region")
         };